chore: rename libs to lib, update eslintrc, and move internal
