+++ ui/lib/analysis/useOfBasicAuth.ts
