+++ ui/lib/internal/index.ts
@@ -9,4 +9,4 @@
    See the License for the specific language governing permissions and
    limitations under the License. */
 
-export { default as logger } from "./logger";
+export { default as logger } from './logger';
