+++ ui/pages/api/analyses.ts
@@ -1,11 +1,7 @@
 import { NextApiRequest, NextApiResponse } from 'next';
-import {
-  Analysis,
-  AnalysisFunction,
-  runAllAnalyses,
-} from '../../libs/findings';
-import { runner as reusedAuthentication } from '../../libs/analysis/reused-authentication';
-import { useOfBasicAuth } from '../../libs/analysis/useOfBasicAuth';
+import { Analysis, AnalysisFunction, runAllAnalyses } from '../../lib/findings';
+import { runner as reusedAuthentication } from '../../lib/analysis/reused-authentication';
+import { useOfBasicAuth } from '../../lib/analysis/useOfBasicAuth';
 
 const analysisFunctions: AnalysisFunction[] = [
   reusedAuthentication,
